Skybox creation
Decided to make a skybox with shaders to make the world look nice. I used visual shaders and ShaderLib to achieve this. Ended up with something pretty cool but the twirling effect will probably cause motion sickness in VR. Duplicated and am working on altering it to be a little less nauseating